Reflex/own/requirements.txt
reflex==0.4.3
python-dotenv==1.0.1    
supabase
configcat-client